Extract API key provisioning to its own method
